feat: moving columns definitions out of component, experimental sorting in keyword table
