+++ admin/src/components/TableViewHeader.jsx
@@ -1,31 +1,17 @@
 import { useState } from 'react';
 import { useI18n } from '@wordpress/react-i18n';
-import { useQueryClient, useMutation } from '@tanstack/react-query';
-import { deleteAll } from '../api/deleteTableData';
 
-import Button from '../elements/Button';
 import SimpleButton from '../elements/SimpleButton';
 
 import '../assets/styles/components/_TableViewHeader.scss';
 
-export default function TableViewHeader( { tableMenu, slug, activeMenu } ) {
+export default function TableViewHeader( { tableMenu, activeMenu } ) {
 	const { __ } = useI18n();
 	const [ active, setActive ] = useState( 'overview' );
-	const queryClient = useQueryClient();
-
-	const handleDelete = useMutation( {
-		mutationFn: () => {
-			return deleteAll( slug );
-		},
-		onSuccess: () => {
-			queryClient.invalidateQueries( [ slug ] );
-		},
-	} );
 
 	const menuItems = new Map( [
 		[ 'overview', __( 'Overview' ) ],
 		[ 'settings', __( 'Settings' ) ],
-		// [ 'importexport', __( 'Import/Export' ) ],
 	] );
 
 	const handleMenu = ( menukey ) => {
@@ -69,9 +55,6 @@ export default function TableViewHeader( { tableMenu, slug, activeMenu } ) {
 				>
 					{ menuItems.get( 'settings' ) }
 				</SimpleButton>
-			</div>
-			<div className="urlslab-tableView-headerBottom">
-				<Button onClick={ () => handleDelete.mutate() }>{ __( 'Delete All' ) }</Button>
 			</div>
 		</div>
 	);
